+++ open-codegen/opengen/constraints/cartesian.py
@@ -4,7 +4,7 @@ from typing import List
 
 class CartesianProduct(constraint.Constraint):
 
-    def __init__(self, dimension: int, segments: List[int], constraints: List[constraint.Constraint]):
+    def __init__(self, segments: List[int], constraints: List[constraint.Constraint]):
         """
         Construct a Cartesian product of constraints by providing a list of sets
         and their dimensions as follows: an n-dimensional vector x can be partitioned
@@ -19,30 +19,53 @@ class CartesianProduct(constraint.Constraint):
         > rect = og.constraints.Rectangle(xmin=[-1,-2,-3], xmax=[0, 10, -1])
         > free = og.constraints.NoConstraints()
         > segment_ids = [1, 4, 9]
-        > dim = 10
+        > my_set = og.constraints.CartesianProduct(segment_ids, [ball, rect])
 
-        :param dimension: dimension of the Cartesian product
         :param segments: ids of segments
         :param constraints: list of sets
         """
-        self.__dimension = dimension
+
+        if any([segments[i] >= segments[i+1] for i in range(len(segments)-1)]):
+            raise ValueError("segments should be a list of integers in strictly ascending order")
+
+        if segments[0] < 0:
+            raise ValueError("the first element of segment must be a positive integer")
+
         self.__segments = segments
         self.__constraints = constraints
 
     @property
     def constraints(self):
+        """
+
+        :return: list of constraints comprising the current instance of CartesianProduct
+        """
         return self.__constraints
 
     @property
     def segments(self):
+        """
+        :return: list of segments
+        """
         return self.__segments
 
     def segment_dimension(self, i):
+        """
+        Dimension of segment i
+        :param i: index of segment (starts at 0)
+        :return: dimension of i-th index
+        """
         if i == 0:
             return self.__segments[0] + 1
         return self.__segments[i] - self.__segments[i-1]
 
     def distance_squared(self, u):
+        """
+        Squared distance of given vector, u, from the current instance of
+        CartesianProduct
+        :param u: vector u
+        :return: squared distance (float)
+        """
         squared_distance = 0.0
         num_segments = len(self.__segments)
         idx_previous = -1